move deprecated cmd in independent file
